Simplify the module import order related to `plain_result/try_catch`
