+++ packages/api_tests/__tests__/plain_result/try_catch_async/try_catch_with_ensure_error_async/producer_is_normal_fn.test.mjs
@@ -1,7 +1,7 @@
 import test from 'ava';
 
-import { tryCatchIntoResultWithEnsureErrorAsync } from 'option-t/plain_result/experimental/try_catch_async';
 import { isOk, isErr, unwrapOk, unwrapErr } from 'option-t/plain_result/result';
+import { tryCatchIntoResultWithEnsureErrorAsync } from 'option-t/plain_result/try_catch_async';
 import { getCrossRealmErrorConstructor } from '../../../cross_realm_error_helper.mjs';
 
 test('output=Ok(T): producer is normal fn', async (t) => {
