--- packages/option-t/src/plain_result/experimental/try_catch.ts
@@ -1,19 +1,0 @@
-import type { ProducerFn } from '../../internal/function.js';
-import { wrapWithNewErrorIfCausalIsUnknown } from '../internal/unknown_causal_carrier.js';
-import { mapErrForResult } from '../map_err.js';
-import type { Result } from '../result.js';
-import { tryCatchIntoResult } from '../try_catch.js';
-
-/**
- *  @experimental
- *  We might change this without any breaking changes.
- *  See https://github.com/option-t/option-t/issues/2295
- */
-export function tryCatchIntoResultWithEnsureError<T>(producer: ProducerFn<T>): Result<T, Error> {
-    const result = tryCatchIntoResult(producer);
-    const mapped: Result<T, Error> = mapErrForResult<T, unknown, Error>(
-        result,
-        wrapWithNewErrorIfCausalIsUnknown,
-    );
-    return mapped;
-}
