+++ packages/option-t/src/plain_result/try_catch_async.ts
@@ -1,6 +1,7 @@
 import { assertIsErrorInstance } from '../internal/assert.js';
 import type { AsyncProducerFn } from '../internal/function.js';
 import { ERR_MSG_THROWN_VALUE_IS_NOT_BUILTIN_ERROR_INSTANCE } from './internal/error_message.js';
+import { wrapWithNewErrorIfCausalIsUnknown } from './internal/unknown_causal_carrier.js';
 import { mapErrForResult } from './map_err.js';
 import { type Result, createOk, createErr } from './result.js';
 
@@ -26,7 +27,21 @@ export async function tryCatchIntoResultAsync<T>(
     return wrapped;
 }
 
-export { tryCatchIntoResultWithEnsureErrorAsync } from './experimental/try_catch_async.js';
+/**
+ *  @experimental
+ *  We might change this without any breaking changes.
+ *  See https://github.com/option-t/option-t/issues/2296
+ */
+export async function tryCatchIntoResultWithEnsureErrorAsync<T>(
+    producer: AsyncProducerFn<T>,
+): Promise<Result<T, Error>> {
+    const result = await tryCatchIntoResultAsync(producer);
+    const mapped: Result<T, Error> = mapErrForResult<T, unknown, Error>(
+        result,
+        wrapWithNewErrorIfCausalIsUnknown,
+    );
+    return mapped;
+}
 
 /**
  *  - This function converts the returend value from _producer_ into `Ok(TValue)`.
